add data baru judul skripsi dan dospem

+++ app/src/main/java/com/example/pertemuan15/repo/NetworkRepositoryMhs.kt
@@ -1,9 +1,10 @@
 package com.example.pertemuan15.repo
 
+import android.util.Log
 import com.example.pertemuan15.model.Mahasiswa
 import com.google.firebase.firestore.FirebaseFirestore
+import com.google.firebase.firestore.FirebaseFirestoreException
 import com.google.firebase.firestore.Query
-import com.google.firebase.firestore.toObject
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.callbackFlow
@@ -41,24 +42,45 @@ class NetworkRepositoryMhs (
             .document(nim)
             .addSnapshotListener { value, error ->
                 if (value != null) {
-                    val mhs = value.toObject(Mahasiswa::class.java)!!
-                    trySend(mhs)
+                    val mhs = value.toObject(Mahasiswa::class.java)
+                    if (mhs != null) {
+                        trySend(mhs)
+                    } else {
+                        trySend(Mahasiswa()) // Default jika data kosong
+                    }
+                } else if (error != null) {
+                    close(error)
                 }
             }
+        awaitClose()
     }
 
-    override suspend fun deleteMhs(mahasiswa: Mahasiswa) {
+
+    override suspend fun deleteMhs(nim: String) {
+        if (nim.isBlank()) {
+            Log.e("RepositoryMhs", "NIM tidak boleh kosong.")
+            throw IllegalArgumentException("NIM tidak boleh kosong.")
+        }
         try {
             firestore.collection("Mahasiswa")
-                .document(mahasiswa.nim)
+                .document(nim)
                 .delete()
                 .await()
+
+            Log.d("RepositoryMhs", "Berhasil menghapus data mahasiswa dengan NIM: $nim")
+        } catch (e: FirebaseFirestoreException) {
+            Log.e("RepositoryMhs", "Kesalahan Firebase saat menghapus mahasiswa: ${e.message}")
+            throw FirebaseFirestoreException(
+                "Gagal menghapus data mahasiswa dari Firestore: ${e.message}",
+                e.code
+            )
         } catch (e: Exception) {
-            throw Exception ("Gagal menghapus data mahasiswa: ${e.message}")
+            Log.e("RepositoryMhs", "Kesalahan saat menghapus mahasiswa: ${e.message}")
+            throw Exception("Gagal menghapus mahasiswa: ${e.message}")
         }
     }
 
-    override suspend fun updateMhs(mahasiswa: Mahasiswa) {
+    override suspend fun updateMhs(mahasiswa: Mahasiswa, nim: String) {
         try {
             firestore.collection("Mahasiswa")
                 .document(mahasiswa.nim)
